requirements.txt
flask
tensorflow
numpy
openpyxl
keras
pillow
opencv-python
opencv-python-headless
PyQt5
